feat: use new module builder
