+++ .editorconfig
@@ -1,4 +1,3 @@
-# editorconfig.org
 root = true
 
 [*]
