--- .eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  root: true,
-  extends: [
-    '@nuxtjs/eslint-config-typescript'
-  ]
-}
