+++ .github/workflows/ci-dev.yml
@@ -38,8 +38,5 @@ jobs:
         if: steps.cache.outputs.cache-hit != 'true'
         run: yarn
 
-      - name: Lint
-        run: yarn lint
-
-      - name: Build
-        run: yarn build
+      - name: Prepack
+        run: yarn prepack
