--- example/.env.example
@@ -1,2 +1,0 @@
-ALGOLIA_APPLICATION_ID=
-ALGOLIA_SEARCH_API_KEY=
