--- example/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "./.nuxt/tsconfig.json",
-  "compilerOptions": {
-    "types": ["@nuxt-commerce/algolia"]
-  }
-}
