+++ tsconfig.json
@@ -1,23 +1,3 @@
 {
-  "compilerOptions": {
-    "allowJs": true,
-    "baseUrl": ".",
-    "strict": true,
-    "target": "ESNext",
-    "module": "ESNext",
-    "lib": ["ESNext", "DOM"],
-    "esModuleInterop": true,
-    "moduleResolution": "node",
-    "declaration": false,
-    "skipLibCheck": true,
-    "resolveJsonModule": true,
-    "allowSyntheticDefaultImports": true,
-    "types": [
-      "node",
-      "@nuxt/kit"
-    ],
-    "paths": {
-      "#app": ["./node_modules/nuxt3/dist/app"]
-    }
-  }
+  "extends": "./playground/.nuxt/tsconfig.json"
 }
